결제 클래스 임시 완성2
